Move DeckPreview outside cardPanel directory
