+++ client/src/components/lobby/deckPreview/style.scss
